Fix rbind_pages issue with named list argument (#326)
* Fix rbind_pages issue with named list argument

`rbind_pages` failed for a named list argument where the argument had a `data.frame` column. In this case, the default behavior of `unlist` is to concatenate the list names with the `data.frame` names and these no longer can be used to index the `data.frame`.

* Fix rbind_pages issue with named list argument

`rbind_pages` failed for a named list argument where the argument had a `data.frame` column. In this case, the default behavior of `unlist` is to concatenate the list names with the `data.frame` names and these no longer can be used to index the `data.frame`.